Create MemeNameGenerator
I wanted to build a name ranomizer as a test of knowlede. I like the Benedict Cumberbatch meme. But the uses here are many, Change the promts and you can get a DnD character name. Make up names for short writing promts, or even plan a silly day out by randomizing your stops for the day.